requirements.txt

pillow
pytesseract
langchain==0.3.21
langchain_community==0.3.20
sentence-transformers==3.4.1
langchain_chroma==0.2.2
langchain-ollama==0.3.0
pypdf==5.4.0
python-dotenv==1.1.0
langchainhub==0.1.21
fastapi ==0.115.12
pydantic== 2.11.4
uvicorn== 0.34.2
fastapi
uvicorn
Pillow
tabulate
keybert
chromadb